chore: rename query key and file transfer
