+++ src/constants/queryKeys.ts
@@ -1,10 +1,14 @@
 const QueryKeys: Record<string, string> = {
+  USER: 'user',
   // Lists & Ideas
   LIST: 'list',
   LISTS: 'lists',
-  USER: 'user',
-  IDEA: 'idea',
   CATEGORIES: 'categories',
+  IDEA: 'idea',
+  IDEAS: 'ideas',
+  ORDER_IDEAS: 'orderIdeas',
+  INFINITE_IDEA: 'infiniteIdea',
+
   OFFICIAL_COLLECTIONS: 'officialCollections',
   LATEST_LIST_GROUPS: 'latestListGroups',
   // Relation
