+++ src/hooks/queries/useCategories.ts
@@ -1,7 +1,7 @@
-import ApiPath from '@/config/apiPath';
+import ApiPath from '@/constants/apiPath';
+import QueryKeys from '@/constants/queryKeys';
 import axios from '@/lib/axios';
 import { Category } from '@/types/List';
-import QueryKeys from '@/config/queryKeys';
 import { IResponse } from '@/types/response';
 import { useQuery } from '@tanstack/react-query';
